+++ src/SharpLink.Server/GlobalUsings.cs
@@ -16,3 +16,5 @@ global using SharpLink.Sdk;
 global using System.IO.Pipes;
 global using System.Net;
 global using System.Net.Sockets;
+global using System.Net.Security;
+global using System.Security.Authentication;